Move all users related files to users module folder
