+++ controllers/dashboard.controller.js
@@ -1,6 +1,6 @@
 const asyncHandler = require('express-async-handler');
 const Report = require('../models/report.model');
-const User = require('../models/user.model');
+const User = require('../modules/users/user.model');
 const Order = require('../modules/orders/order.model');
 const Player = require('../modules/players/player.model');
 const Club = require('../modules/clubs/club.model');
