+++ middleware/__tests__/auth.test.js
@@ -1,6 +1,6 @@
 const { protect, authorize } = require('../auth');
 const { buildUser, buildReq, buildRes, buildNext } = require('../../test/utils');
-const User = require('../../models/user.model');
+const User = require('../../modules/users/user.model');
 
 describe('route protection middleware', () => {
   it('should throw a 401 error if no token provided', () => {
