+++ scripts/seeder.js
@@ -5,7 +5,7 @@ const dotenv = require('dotenv');
 
 dotenv.config({ path: './config/config.env' });
 
-const User = require('../models/user.model');
+const User = require('../modules/users/user.model');
 const Club = require('../modules/clubs/club.model');
 const Player = require('../modules/players/player.model');
 const Order = require('../modules/orders/order.model');
