+++ services/db.service.js
@@ -1,6 +1,6 @@
 const Player = require('../modules/players/player.model');
 const Club = require('../modules/clubs/club.model');
-const User = require('../models/user.model');
+const User = require('../modules/users/user.model');
 const Order = require('../modules/orders/order.model');
 const Report = require('../models/report.model');
